[gazelle] Generalize scc collapse to all JVM languages; rename kotlinscc -> scc
The collapse is only Kotlin-specific for the `internal`-coupling edge; the
import-cycle edge, the graph, and the per-group emit are all language-agnostic.
A pure-Java module with circular package deps -- which javac accepts but
Bazel's acyclic target graph rejects -- collapses exactly the cyclic packages
into one java_library and leaves the rest fine-grained, a strictly finer result
than `module` granularity.

Rename the `kotlinscc` package to `scc` and reword its docs. Add fixtures for
all three cases: pure Java (Kotlin disabled), the existing Kotlin module, and a
mixed Java<->Kotlin cycle that collapses into one kt_jvm_library.
