+++ java/gazelle/BUILD.bazel
@@ -24,9 +24,9 @@ go_library(
         "//java/gazelle/private/java_export_index",
         "//java/gazelle/private/javaparser",
         "//java/gazelle/private/kotlin",
-        "//java/gazelle/private/kotlinscc",
         "//java/gazelle/private/logconfig",
         "//java/gazelle/private/maven",
+        "//java/gazelle/private/scc",
         "//java/gazelle/private/sorted_multiset",
         "//java/gazelle/private/sorted_set",
         "//java/gazelle/private/types",
