+++ java/gazelle/private/scc/scc_test.go
@@ -1,4 +1,4 @@
-package kotlinscc
+package scc
 
 import (
 	"reflect"
