+++ java/gazelle/testdata/java_module_scc/maven_install.json
@@ -1,0 +1,1 @@
+{"version": "2"}
