+++ java/gazelle/testdata/java_module_scc/src/main/java/com/example/a/A.java
@@ -1,0 +1,9 @@
+package com.example.a;
+
+import com.example.b.B;
+
+public class A {
+  public B makeB() {
+    return new B();
+  }
+}
