+++ java/gazelle/testdata/java_module_scc/src/main/java/com/example/b/B.java
@@ -1,0 +1,9 @@
+package com.example.b;
+
+import com.example.a.A;
+
+public class B {
+  public A makeA() {
+    return new A();
+  }
+}
