+++ java/gazelle/testdata/java_module_scc/src/main/java/com/example/solo/Solo.java
@@ -1,0 +1,9 @@
+package com.example.solo;
+
+import com.example.a.A;
+
+public class Solo {
+  public A useA() {
+    return new A();
+  }
+}
