+++ java/gazelle/testdata/mixed_module_scc/src/main/java/com/example/BUILD.in
@@ -1,0 +1,1 @@
+# gazelle:java_module_granularity scc
