+++ java/gazelle/testdata/mixed_module_scc/src/main/java/com/example/BUILD.out
@@ -1,0 +1,12 @@
+load("@rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+# gazelle:java_module_granularity scc
+
+kt_jvm_library(
+    name = "j",
+    srcs = [
+        "j/J.java",
+        "k/K.kt",
+    ],
+    visibility = ["//:__subpackages__"],
+)
