+++ java/gazelle/testdata/mixed_module_scc/src/main/java/com/example/j/J.java
@@ -1,0 +1,9 @@
+package com.example.j;
+
+import com.example.k.K;
+
+public class J {
+  public K makeK() {
+    return new K();
+  }
+}
