+++ java/gazelle/testdata/mixed_module_scc/src/main/java/com/example/k/K.kt
@@ -1,0 +1,7 @@
+package com.example.k
+
+import com.example.j.J
+
+class K {
+  fun makeJ(): J = J()
+}
